requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
alembic=1.15.1=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=24.3.0=pypi_0
blas=1.0=mkl
bottleneck=1.3.7=pypi_0
brotli=1.0.9=pypi_0
brotli-bin=1.0.9=h5eee18b_9
brotli-python=1.0.9=py310h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.1.31=pypi_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
class-resolver=0.5.4=pypi_0
click=8.1.7=pypi_0
click-default-group=1.2.4=pypi_0
colorlog=6.9.0=pypi_0
comm=0.2.1=pypi_0
contourpy=1.3.1=pypi_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cuda-version=12.6=3
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h9c0eb46_1
dataclasses-json=0.6.7=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pyhd3eb1b0_0
dgl=2.4.0+cu118=pypi_0
docdata=0.0.4=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.6.4=h6a678d5_0
fair-esm=2.0.0=pypi_0
fastjsonschema=2.20.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=pypi_0
fontconfig=2.14.1=hef1e5e3_0
fonttools=4.51.0=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2025.3.0=pypi_0
giflib=5.2.2=h5eee18b_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=pypi_0
gnutls=3.6.15=he1e5248_0
greenlet=3.1.1=pypi_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
icu=58.2=he6710b0_3
idna=3.7=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.29.5=pypi_0
ipython=8.27.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_3
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
jupyter-client=8.6.0=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter_client=8.6.0=py310h06a4308_0
jupyter_core=5.7.2=py310h06a4308_0
kiwisolver=1.4.4=pypi_0
krb5=1.20.1=h568e23c_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h12ee557_0
libbrotlicommon=1.0.9=h5eee18b_9
libbrotlidec=1.0.9=h5eee18b_9
libbrotlienc=1.0.9=h5eee18b_9
libclang=14.0.6=default_hc6dbbc7_2
libclang13=14.0.6=default_he11475f_2
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.11.1.6=0
libcups=2.4.2=ha637b67_0
libcurand=10.3.7.77=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libedit=3.1.20230828=h5eee18b_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.14=0
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libllvm14=14.0.6=hef93074_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libpq=12.15=h37d81fd_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.1.0=hecacb30_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=0
libwebp=1.2.0=h89dd481_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
mako=1.3.9=pypi_0
markupsafe=2.1.3=pypi_0
marshmallow=3.26.1=pypi_0
matplotlib=3.10.0=pypi_0
matplotlib-base=3.10.0=py310hbfdbfaf_0
matplotlib-inline=0.1.6=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-fft=1.3.1=pypi_0
mkl-random=1.2.2=pypi_0
mkl-service=2.4.0=pypi_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
more-click=0.1.2=pypi_0
more-itertools=10.6.0=pypi_0
mpc=1.0.3=0
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=pypi_0
mypy-extensions=1.0.0=pypi_0
mysql=5.7.24=he378463_2
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=3.3=pypi_0
numexpr=2.8.4=pypi_0
numpy=1.24.3=pypi_0
numpy-base=1.24.3=py310h8e6c178_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1w=h7f8727e_0
optuna=4.2.1=pypi_0
packaging=24.1=pypi_0
pandas=2.2.3=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pillow=9.3.0=pypi_0
pip=24.2=pypi_0
platformdirs=3.10.0=pypi_0
plotly=5.24.1=pypi_0
ply=3.11=pypi_0
pooch=1.7.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
prompt_toolkit=3.0.43=hd3eb1b0_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pygments=2.15.1=pypi_0
pykeen=1.11.0=pypi_0
pyparsing=3.2.0=pypi_0
pyqt=5.15.10=py310h6a678d5_0
pyqt5=5.15.10=pypi_0
pyqt5-sip=12.13.0=pypi_0
pysocks=1.7.1=pypi_0
pystow=0.7.0=pypi_0
python=3.10.0=h12debd9_5
python-dateutil=2.9.0.post0=pypi_0
python-fastjsonschema=2.20.0=py310h06a4308_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.2.2=py3.10_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_6
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=25.1.2=pypi_0
qt-main=5.15.2=h5b8104b_9
readline=8.2=h5eee18b_0
referencing=0.30.2=pypi_0
requests=2.32.3=pypi_0
rpds-py=0.22.3=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.13.2=pypi_0
setuptools=75.1.0=pypi_0
sip=6.7.12=pypi_0
six=1.16.0=pyhd3eb1b0_1
sqlalchemy=2.0.38=pypi_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.13.2=pypi_0
tabulate=0.9.0=pypi_0
tenacity=9.0.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.11=h1ccaba5_0
tomli=2.0.1=pypi_0
torch=2.2.2=pypi_0
torch-max-mem=0.1.3=pypi_0
torch-ppr=0.0.8=pypi_0
torchaudio=2.2.2=pypi_0
torchtriton=2.2.0=py310
torchvision=0.17.2=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.66.5=pypi_0
traitlets=5.14.3=pypi_0
triton=2.2.0=pypi_0
typing-extensions=4.12.2=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.11.0=py310h06a4308_0
tzdata=2024b=h04d1e81_0
unicodedata2=15.1.0=pypi_0
urllib3=2.2.3=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.44.0=pypi_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zeromq=4.3.5=h6a678d5_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
